+++ src/main/java/com/bitbox/bitboxserver/dao/SupplierDAO.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface SupplierDAO extends CrudRepository<Supplier, Long> {
-    public Supplier findBySupplierCode(int code);
+    public Supplier findBySupplierCode(Long code);
 }
