+++ src/main/java/com/bitbox/bitboxserver/dao/UserDAO.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface UserDAO extends CrudRepository<User, Long> {
-    public User findByUserCode(int code);
-    public void deleteByUserCode(int code);
+    public User findByUserCode(Long code);
+    public void deleteByUserCode(Long code);
 }
