+++ src/main/java/com/bitbox/bitboxserver/service/IUserService.java
@@ -7,8 +7,8 @@ import java.util.List;
 
 public interface IUserService {
     public List<UserDTO> findAllUsers();
-    public UserDTO findByUserCode(int code);
+    public UserDTO findByUserCode(Long code);
     public void createUser(UserDTO userDTO);
-    public void deleteUser(int code);
-    public ResponseEntity<UserDTO> updateUser(int code, UserDTO userDTO);
+    public void deleteUser(Long code);
+    public ResponseEntity<UserDTO> updateUser(Long code, UserDTO userDTO);
 }
